Initial commit
We added the main.py file which is the main script file for the application. The file displays output and prints text on the console screen without colors, and that colored output feature needs to be added soon. Also, the proper README.md file is not made, it will be updated soon.